ui draft for spotify synchronization
