+++ webapp/src/main/resources/static/css/common.css
@@ -154,6 +154,7 @@
 .material-icons.md-24 { font-size: 24px; }
 .material-icons.md-30 { font-size: 30px; }
 .material-icons.md-36 { font-size: 36px; }
+.material-icons.md-42 { font-size: 42px; }
 .material-icons.md-48 { font-size: 48px; }
 
 .material-icons.md-dark { color: rgba(0, 0, 0, 0.54); }
