+++ webapp/src/main/resources/static/ts/src/bundles/my-artists.ts
@@ -5,7 +5,6 @@ import {ToastService} from "../service/toast-service";
 import {AlertService} from "../service/alert-service";
 import {LoadingIndicatorService} from "../service/loading-indicator-service";
 import {DateFormatService} from "../service/date-format-service";
-import {SpotifyRestClient} from "../clients/spotify-rest-client";
 import {UrlService} from "../service/url-service";
 
 const urlService = new UrlService();
@@ -14,9 +13,8 @@ const alertService = new AlertService();
 const loadingIndicatorService = new LoadingIndicatorService();
 const artistsRestClient = new ArtistsRestClient(urlService, toastService);
 const followArtistService = new FollowArtistService(artistsRestClient, toastService);
-const spotifyRestClient = new SpotifyRestClient(toastService);
 const dateFormatService = new DateFormatService();
-const myArtistsRenderService = new MyArtistsRenderService(followArtistService, dateFormatService, alertService, loadingIndicatorService, spotifyRestClient, toastService);
+const myArtistsRenderService = new MyArtistsRenderService(followArtistService, dateFormatService, alertService, loadingIndicatorService);
 
 const response = artistsRestClient.fetchMyArtists();
 myArtistsRenderService.render(response);
