--- webapp/src/main/resources/static/ts/src/bundles/settings.ts
@@ -1,9 +1,0 @@
-import {SettingsRenderService} from "../service/settings-render-service";
-import {ToastService} from "../service/toast-service";
-import {SpotifyRestClient} from "../clients/spotify-rest-client";
-
-const toastService = new ToastService();
-const spotifyAuthorizationRestClient = new SpotifyRestClient(toastService);
-const settingsRenderService = new SettingsRenderService(spotifyAuthorizationRestClient);
-
-settingsRenderService.render();
