+++ webapp/src/main/resources/static/ts/src/service/my-artists-render-service.ts
@@ -7,8 +7,6 @@ import {LoadingIndicatorService} from "./loading-indicator-service";
 import {PaginationComponent} from "../components/pagination/pagination-component";
 import {AbstractRenderService} from "./abstract-render-service";
 import {DateFormat, DateFormatService} from "./date-format-service";
-import {SpotifyRestClient} from "../clients/spotify-rest-client";
-import {ToastService} from "./toast-service";
 
 export class MyArtistsRenderService extends AbstractRenderService<MyArtistsResponse> {
 
@@ -18,17 +16,12 @@ export class MyArtistsRenderService extends AbstractRenderService<MyArtistsRespo
     private readonly dateFormatService: DateFormatService;
     private readonly paginationComponent: PaginationComponent;
     private readonly artistTemplateElement: HTMLTemplateElement;
-    private readonly spotifyRestClient: SpotifyRestClient;
-    private readonly toastService: ToastService;
     private rowElement?: HTMLDivElement;
 
-    constructor(followArtistService: FollowArtistService, dateFormatService: DateFormatService, alertService: AlertService, loadingIndicatorService: LoadingIndicatorService,
-      spotifyRestClient: SpotifyRestClient, toastService: ToastService) {
+    constructor(followArtistService: FollowArtistService, dateFormatService: DateFormatService, alertService: AlertService, loadingIndicatorService: LoadingIndicatorService) {
         super(alertService, loadingIndicatorService);
         this.followArtistService = followArtistService;
         this.dateFormatService = dateFormatService;
-        this.spotifyRestClient = spotifyRestClient;
-        this.toastService = toastService;
         this.paginationComponent = new PaginationComponent();
         this.artistTemplateElement = document.getElementById("artist-card")! as HTMLTemplateElement;
     }
@@ -38,8 +31,6 @@ export class MyArtistsRenderService extends AbstractRenderService<MyArtistsRespo
     }
 
     protected onRendering(response: MyArtistsResponse): void {
-        this.addEventListener();
-
         if (response.myArtists.length === 0 && response.pagination.currentPage === 1) {
             const message = '<h3 class="h5">Currently you do not follow any artist.</h3>Start a search for your favorite artists right now.';
             const infoMessage = this.alertService.renderInfoAlert(message, false);
@@ -110,15 +101,5 @@ export class MyArtistsRenderService extends AbstractRenderService<MyArtistsRespo
     private attachPagination(paginationData: Pagination): void {
         const paginationList = this.paginationComponent.render(paginationData);
         this.hostElement.insertAdjacentElement("beforeend", paginationList);
-    }
-
-    private addEventListener(): void {
-        const importButton = document.getElementById("spotify-import-button")! as HTMLButtonElement;
-        importButton.addEventListener("click", this.doSpotifyArtistImport.bind(this));
-    }
-
-    private doSpotifyArtistImport(): void {
-        const importResponse = this.spotifyRestClient.importArtists();
-        importResponse.then(response => this.toastService.createInfoToast("Successfully imported " + response.artists.length + " artists!"));
     }
 }
