+++ webapp/src/main/resources/static/ts/webpack.config.prod.js
@@ -5,10 +5,10 @@ module.exports = {
     mode: "production",
     entry: {
         search: "./src/bundles/search.ts",
-        myartists: "./src/bundles/my-artists.ts",
+        my_artists: "./src/bundles/my-artists.ts",
         releases: "./src/bundles/releases.ts",
         homepage: "./src/bundles/homepage.ts",
-        settings: "./src/bundles/settings.ts"
+        spotify_synchronization: "./src/bundles/spotify-synchronization.ts"
     },
     output: {
         filename: "[name].bundle.js",
