+++ webapp/src/main/resources/templates/frontend/fragments/layout.html
@@ -86,11 +86,11 @@
       <div class="container p-3">
         <div class="row float-right">
           <ul class="list-inline">
-            <li class="list-inline-item"><a class="hyperlink" href="#" th:href="@{https://github.com/MetalDetectorRocks?tab=repositories}" target="_blank">Find us on GitHub</a></li>
-            <li class="list-inline-item"><a class="hyperlink" href="#" th:href="@{https://github.com/MetalDetectorRocks/metal-detector-main/issues/new?assignees=&labels=&template=feature_request.md&title=}" target="_blank">Request a feature</a></li>
-            <li class="list-inline-item"><a class="hyperlink" href="#" th:href="@{https://github.com/MetalDetectorRocks/metal-detector-main/issues/new?assignees=&labels=&template=bug_report.md&title=}" target="_blank">Report an issue</a></li>
-            <li class="list-inline-item"><a class="hyperlink" href="#" target="_blank">Spend a beer</a></li>
-            <li class="list-inline-item"><a class="hyperlink" href="#" th:href="@{/imprint}">Imprint</a></li>
+            <li class="list-inline-item"><a class="dark-hyperlink" href="#" th:href="@{https://github.com/MetalDetectorRocks?tab=repositories}" target="_blank">Find us on GitHub</a></li>
+            <li class="list-inline-item"><a class="dark-hyperlink" href="#" th:href="@{https://github.com/MetalDetectorRocks/metal-detector-main/issues/new?assignees=&labels=&template=feature_request.md&title=}" target="_blank">Request a feature</a></li>
+            <li class="list-inline-item"><a class="dark-hyperlink" href="#" th:href="@{https://github.com/MetalDetectorRocks/metal-detector-main/issues/new?assignees=&labels=&template=bug_report.md&title=}" target="_blank">Report an issue</a></li>
+            <li class="list-inline-item"><a class="dark-hyperlink" href="#" target="_blank">Spend a beer</a></li>
+            <li class="list-inline-item"><a class="dark-hyperlink" href="#" th:href="@{/imprint}">Imprint</a></li>
           </ul>
         </div>
       </div>
