+++ webapp/src/main/resources/templates/frontend/my-artists.html
@@ -5,7 +5,7 @@
       lang="en">
 <head>
     <title>My Artists</title>
-    <script type="module" th:src="@{/ts/dist/myartists.bundle.js}" src="#"></script>
+    <script type="module" th:src="@{/ts/dist/my_artists.bundle.js}" src="#"></script>
 </head>
 <body>
     <section layout:fragment="content">
@@ -24,9 +24,7 @@
         </template>
 
         <h1 class="h4">My Artists</h1>
-        <div id="artists-container" class="mb-3">
-            <button type="button" class="btn btn-dark btn-sm mt-2 mb-3" id="spotify-import-button">Import from Spotify</button>
-        </div>
+        <div id="artists-container" class="mb-3"></div>
     </section>
 </body>
 </html>
