+++ webapp/src/main/resources/templates/frontend/spotify-synchronization.html
@@ -5,16 +5,115 @@
       lang="en">
 <head>
     <title>Spotify Synchronization</title>
-    <script type="module" th:src="@{/ts/dist/settings.bundle.js}" src="#"></script>
+    <script type="module" th:src="@{/ts/dist/spotify_synchronization.bundle.js}" src="#"></script>
 </head>
 <body>
     <section layout:fragment="content">
+        <template>
+            <div class="card mt-2 spotify-synchro-card pointer">
+                <div class="row">
+                    <div class="col-md-auto">
+                        <img class="spotify-synchro-img" src="" alt="" id="thumb">
+                    </div>
+                    <div class="col">
+                        <p class="h5 mt-3" id="artist-name"></p>
+                        <p id="artist-info"></p>
+                    </div>
+                    <div class="col col-sm-2 text-center spotify-synchro-checkbox">
+                        <span class="material-icons md-42">check_box</span>
+                    </div>
+                </div>
+            </div>
+        </template>
+
         <div class="row">
             <div th:replace="frontend/fragments/settings-menu :: settings-menu"></div>
             <div class="col-8">
                 <h2 class="h5 mt-3">Spotify Synchronization</h2>
                 <hr class="border-secondary"/>
-                <button type="button" class="btn btn-success" id="authorization-button">Authorize Spotify</button>
+                <button type="button" class="btn btn-success" id="spotify-connect-button">Connect with Spotify</button>
+                <button type="button" class="btn btn-success" id="spotify-synchronize-button">Synchronize artists</button>
+                <div class="mt-3 text-right invisible" id="foo">
+                    <a class="light-hyperlink mr-2" href="#">Select all</a> | <a class="light-hyperlink ml-2" href="#">Deselect all</a>
+                </div>
+                <div id="artists-container">
+<!--                    <div class="card mt-2 spotify-synchro-card pointer">-->
+<!--                        <div class="row">-->
+<!--                            <div class="col-md-auto">-->
+<!--                                <img class="spotify-synchro-img img-inactive" src="/images/dummy/abbath.jpg" alt="" id="artist-thumb">-->
+<!--                            </div>-->
+<!--                            <div class="col">-->
+<!--                                <p class="h5 mt-3 text-muted">Abbath</p>-->
+<!--                                <p class="text-muted">-->
+<!--                                    Black Metal, Post Black Metal<br />-->
+<!--                                    1.890 followers on Spotify-->
+<!--                                </p>-->
+<!--                            </div>-->
+<!--                            <div class="col col-sm-2 text-center spotify-synchro-checkbox">-->
+<!--                                <span class="material-icons md-42">-->
+<!--                                    check_box_outline_blank-->
+<!--                                </span>-->
+<!--                            </div>-->
+<!--                        </div>-->
+<!--                    </div>-->
+<!--                    <div class="card mt-2 spotify-synchro-card pointer">-->
+<!--                        <div class="row">-->
+<!--                            <div class="col-md-auto">-->
+<!--                                <img class="spotify-synchro-img" src="/images/dummy/alcest.jpg" alt="" id="artist-thumb-2">-->
+<!--                            </div>-->
+<!--                            <div class="col">-->
+<!--                                <p class="h5 mt-3">Alcest</p>-->
+<!--                                <p>-->
+<!--                                    Black Metal, Post Black Metal<br />-->
+<!--                                    1.890 followers on Spotify-->
+<!--                                </p>-->
+<!--                            </div>-->
+<!--                            <div class="col col-sm-2 text-center spotify-synchro-checkbox">-->
+<!--                                <span class="material-icons md-42 md-success">-->
+<!--                                    check_box-->
+<!--                                </span>-->
+<!--                            </div>-->
+<!--                        </div>-->
+<!--                    </div>-->
+<!--                    <div class="card mt-2 spotify-synchro-card pointer">-->
+<!--                        <div class="row">-->
+<!--                            <div class="col-md-auto">-->
+<!--                                <img class="spotify-synchro-img" src="/images/dummy/harakiri-for-the-sky.jpg" alt="" id="artist-thumb-3">-->
+<!--                            </div>-->
+<!--                            <div class="col">-->
+<!--                                <p class="h5 mt-3">Harakiri For The Sky</p>-->
+<!--                                <p>-->
+<!--                                    Black Metal, Post Black Metal<br />-->
+<!--                                    1.890 followers on Spotify-->
+<!--                                </p>-->
+<!--                            </div>-->
+<!--                            <div class="col col-sm-2 text-center spotify-synchro-checkbox">-->
+<!--                                <span class="material-icons md-42 md-success">-->
+<!--                                    check_box-->
+<!--                                </span>-->
+<!--                            </div>-->
+<!--                        </div>-->
+<!--                    </div>-->
+<!--                    <div class="card mt-2 spotify-synchro-card pointer">-->
+<!--                        <div class="row">-->
+<!--                            <div class="col-md-auto">-->
+<!--                                <img class="spotify-synchro-img img-inactive" src="/images/dummy/marduk.jpg" alt="" id="artist-thumb-4">-->
+<!--                            </div>-->
+<!--                            <div class="col">-->
+<!--                                <p class="h5 mt-3 text-muted">Marduk</p>-->
+<!--                                <p class="text-muted">-->
+<!--                                    Black Metal, Post Black Metal<br />-->
+<!--                                    1.890 followers on Spotify-->
+<!--                                </p>-->
+<!--                            </div>-->
+<!--                            <div class="col col-sm-2 text-center spotify-synchro-checkbox">-->
+<!--                                <span class="material-icons md-42">-->
+<!--                                    check_box_outline_blank-->
+<!--                                </span>-->
+<!--                            </div>-->
+<!--                        </div>-->
+<!--                    </div>-->
+                </div>
             </div>
         </div>
     </section>
